+++ sensor/admission-control/manager/k8s_object.go
@@ -1,6 +1,7 @@
 package manager
 
 import (
+	openshift_appsv1 "github.com/openshift/api/apps/v1"
 	"github.com/pkg/errors"
 	"github.com/stackrox/rox/pkg/k8sutil"
 	"github.com/stackrox/rox/pkg/kubernetes"
@@ -36,6 +37,8 @@ func unmarshalK8sObject(gvk metav1.GroupVersionKind, raw []byte) (k8sutil.Object
 		obj = &batchV1beta1.CronJob{}
 	case kubernetes.Job:
 		obj = &batchV1.Job{}
+	case kubernetes.DeploymentConfig:
+		obj = &openshift_appsv1.DeploymentConfig{}
 	default:
 		return nil, errors.Errorf("currently do not recognize kind %q in admission controller", gvk.Kind)
 	}